fix(hues): adjust `Pmenu*Sel` to work with colored completion items
Completion items on Neovim>=0.11 are allowed to have additional
highlighting (through `hl_group` and `kind_hlgroup` fields) which will
be combined with `Pmenu`, `PmenuSel`, and all its narrower variants.

As `PmenuSel` is (almost) an inverse of `Pmenu`, added highlighting will
be readable only combined with one of those. As it is reasonable to
assume completion item highlighting be optimized for `Pmenu` (i.e. as it
is shown in pop-up menu without selection), `PmenuSel` needs adjusting.

The way it is done (and can be done for any highlight group which may be
combined with attributes designed for background with inverse lightness)
is by switching `fg` and `bg` while adding `reverse = true`. It leaves
items without highlighting look the same as earlier, while combined
highlighting will first combine with switched `fg`/`bg` and only then
reverse (for example, resulting in a colored background if added
highlighting defines only foreground).